plugin/src/main/java/com/s6/plugin/feign/adapter/utils/CollectionUtils.java
package com.s6.plugin.feign.adapter.utils;

import java.util.Collection;
import java.util.Map;

/**
 * 集合工具类
 *
 * @author Sean
 */
public class CollectionUtils {
    private CollectionUtils() {

    }

    public static boolean isEmpty(Collection<?> collection) {
        return (collection == null || collection.isEmpty());
    }

    public static boolean isNotEmpty(Collection<?> collection) {
        return !isEmpty(collection);
    }

    /**
     * Return {@code true} if the supplied Map is {@code null} or empty.
     * Otherwise, return {@code false}.
     *
     * @param map the Map to check
     * @return whether the given Map is empty
     */
    public static boolean isEmpty(Map<?, ?> map) {
        return (map == null || map.isEmpty());
    }
}
